[#31] feat:추천페이지 테그 목록 조회 추가

+++ src/main/java/com/team/delightserver/web/dto/response/MachineLearningResultResponse.java
@@ -17,5 +17,5 @@ import java.util.List;
 public class MachineLearningResultResponse {
 
     private List<String> foods;
-    private List<Integer> scores;
+    private List<Double> scores;
 }
